update: 把"短点"功能独立为模块,并测试可用 fix: 资源文件夹应该叫assets

+++ nutzcn-common/src/main/java/net/wendal/nutzbook/core/module/UserProfileModule.java
@@ -200,7 +200,7 @@ public class UserProfileModule extends BaseModule {
 			if (System.currentTimeMillis() - time > 10*60*1000) {
 				return "该验证链接已经超时";
 			}
-			int userId = Integer.parseInt(tmp[0]);
+			long userId = Long.parseLong(tmp[0]);
 			Cnd cnd = Cnd.where("userId", "=", userId).and("email", "=", tmp[1]);
 			int re = dao.update(UserProfile.class, Chain.make("emailChecked", true), cnd);
 			if (re == 1) {
